remove old bulk rename
